+++ src/Psl/File/Internal/AbstractHandleWrapper.php
@@ -79,6 +79,15 @@ abstract class AbstractHandleWrapper implements File\HandleInterface
      * @inheritDoc
      */
     #[Override]
+    public function isClosed(): bool
+    {
+        return $this->handle->isClosed();
+    }
+
+    /**
+     * @inheritDoc
+     */
+    #[Override]
     public function close(): void
     {
         $this->handle->close();