+++ src/Psl/IO/CloseReadWriteStreamHandle.php
@@ -90,6 +90,15 @@ final class CloseReadWriteStreamHandle implements
      * @inheritDoc
      */
     #[Override]
+    public function isClosed(): bool
+    {
+        return $this->handle->isClosed();
+    }
+
+    /**
+     * @inheritDoc
+     */
+    #[Override]
     public function close(): void
     {
         $this->handle->close();