+++ src/Psl/Unix/Internal/Listener.php
@@ -126,6 +126,15 @@ final class Listener implements Unix\ListenerInterface
      * @inheritDoc
      */
     #[Override]
+    public function isClosed(): bool
+    {
+        return null === $this->impl;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    #[Override]
     public function close(): void
     {
         EventLoop::disable($this->watcher);